rename generate and adapt consumption fkt
