--- examples/consumption.json
@@ -1,42 +1,0 @@
-{
-  "CKB": {
-    "energy_type": "Diesel",
-    "unit": "L/100km",
-    "factor": 0.01,
-    "winter": {
-      "fluent": 33.32,
-      "dense": 47.44,
-      "stop&go": 65.24
-    },
-    "summer": {
-      "fluent": 34.01,
-      "dense": 46.48,
-      "sto  p&go": 53.14
-    },
-    "mean": {
-      "fluent": 33.13,
-      "dense": 46.21,
-      "stop&go": 51.76
-    }
-  },
-  "VDL": {
-    "energy_type": "Diesel",
-    "unit": "L/100km",
-    "factor": 0.01,
-    "winter": {
-      "fluent": 33.32,
-      "dense": 47.44,
-      "stop&go": 65.24
-    },
-    "summer": {
-      "fluent": 34.01,
-      "dense": 46.48,
-      "stop&go": 53.14
-    },
-    "no_hc": {
-      "fluent": 33.13,
-      "dense": 46.21,
-      "stop&go": 51.76
-    }
-  }
-}
