+++ examples/consumption_inputs/default_temp_summer_day_09_08_2020.csv
@@ -1,0 +1,25 @@
+Time,Temperature in C
+00:00:00,23.8
+01:00:00,23.8
+02:00:00,22.7
+03:00:00,22.2
+04:00:00,21.1
+05:00:00,20
+06:00:00,20
+07:00:00,21.1
+08:00:00,22.7
+09:00:00,25
+10:00:00,28.8
+11:00:00,31.1
+12:00:00,32.7
+13:00:00,33.8
+14:00:00,36.1
+15:00:00,36.1
+16:00:00,36.1
+17:00:00,33.8
+18:00:00,32.2
+19:00:00,32.2
+20:00:00,27.2
+21:00:00,25
+22:00:00,23.8
+23:00:00,25
