+++ examples/consumption_inputs/default_temp_winter_day_08_01_2018.csv
@@ -1,0 +1,25 @@
+Time,Temperature in C
+00:00:00,-1.1
+01:00:00,-1.1
+02:00:00,-2.2
+03:00:00,-2.7
+04:00:00,-2.7
+05:00:00,-2.7
+06:00:00,-2.7
+07:00:00,-3.8
+08:00:00,-3.8
+09:00:00,-3.8
+10:00:00,-3.8
+11:00:00,-2.7
+12:00:00,-1.1
+13:00:00,1.1
+14:00:00,2.2
+15:00:00,2.2
+16:00:00,1.1
+17:00:00,0
+18:00:00,0
+19:00:00,-1.1
+20:00:00,-1.1
+21:00:00,-1.1
+22:00:00,-1.1
+23:00:00,-1.1
