+++ examples/generate_distributed_charging.cfg
@@ -27,8 +27,8 @@ v2g = false
 # set possible power at grid connection point in kW
 gc_power = 1500
 # define hours of traffic_jam (options are "stop&go" and "dense")
-traffic_jam = {"stop&go": [["07:00", "09:00"], ["15:00", "17:00"]], "dense": [["17:00", "19:00"]]}
-# define preferred charging type, either opp or depot, default: opp
+rush_hour = {"stop&go": [["07:00", "09:00"], ["15:00", "17:00"]], "dense": [["17:00", "19:00"]]}
+# define preferred charging type depot, default None
 preferred_ct = "depot"
