--- examples/heating_consumption.csv
@@ -1,18 +1,0 @@
-fluent,dense,stop&go,temperature,speed
-0.45,0.58,0.75,-2.5,30
-0.422,0.54,0.7,-1.25,24
-0.405,0.5,0.635,0,18
-0.375,0.466,0.59,1.25,0
-0.35,0.425,0.535,2.5,0
-0.32,0.385,0.48,3.75,0
-0.298,0.35,0.433,5,0
-0.275,0.32,0.385,6.25,0
-0.24,0.28,0.35,7.5,0
-0.21,0.24,0.305,8.75,0
-0.18,0.218,0.266,10,0
-0.15,0.178,0.233,11.25,0
-0.12,0.15,0.198,12.5,0
-0.095,0.105,0.175,13.75,0
-0.06,0.078,0.125,15,0
-0.03,0.05,0.095,16.25,0
-0,0.022,0.066,17.5,0
